requirements.txt
openai
Flask
requests
PyGithub
gitpython
